requirement.txt
requests
beautifulsoup4
Pillow
pyglet
